Fixed roles and usernames
Party is assigned when scene changes. Usernames still don't work, idk fuck mirror. President is assigned to the first player in the lobby also. If you try to get the actual players use the NetworkManagerLobby.roomSlots and cast the players to NetworkRoomPlayerLobby. The players should have the correct authority, if they don't idfk blame mirror.